workshop-dlt-ingestion/my-dlt-pipeline/requirements.txt
dlt[duckdb]>=1.22.1
marimo>=0.11.0
ibis-framework[duckdb]>=10.0.0
altair>=5.0.0